refactor(ngcc): move shared code into `DependencyHostBase`
The various dependency hosts had a lot of duplicated code.
This commit refactors them to move this into the base class.
