+++ packages/compiler-cli/ngcc/src/dependencies/dependency_host.ts
@@ -9,7 +9,7 @@ import {AbsoluteFsPath, FileSystem, PathSegment} from '../../../src/ngtsc/file_s
 import {EntryPoint} from '../packages/entry_point';
 import {resolveFileWithPostfixes} from '../utils';
 
-import {ModuleResolver} from './module_resolver';
+import {ModuleResolver, ResolvedDeepImport, ResolvedRelativeModule} from './module_resolver';
 
 export interface DependencyHost {
   collectDependencies(
@@ -65,7 +65,54 @@ export abstract class DependencyHostBase implements DependencyHost {
    * @param alreadySeen A set that is used to track internal dependencies to prevent getting stuck
    * in a circular dependency loop.
    */
-  protected abstract recursivelyCollectDependencies(
+  protected recursivelyCollectDependencies(
       file: AbsoluteFsPath, dependencies: Set<AbsoluteFsPath>, missing: Set<string>,
-      deepImports: Set<AbsoluteFsPath>, alreadySeen: Set<AbsoluteFsPath>): void;
+      deepImports: Set<string>, alreadySeen: Set<AbsoluteFsPath>): void {
+    const fromContents = this.fs.readFile(file);
+    if (this.canSkipFile(fromContents)) {
+      return;
+    }
+    const imports = this.extractImports(file, fromContents);
+    for (const importPath of imports) {
+      const resolved =
+          this.processImport(importPath, file, dependencies, missing, deepImports, alreadySeen);
+      if (!resolved) {
+        missing.add(importPath);
+      }
+    }
+  }
+
+  protected abstract canSkipFile(fileContents: string): boolean;
+  protected abstract extractImports(file: AbsoluteFsPath, fileContents: string): Set<string>;
+
+  /**
+   * Resolve the given `importPath` from `file` and add it to the appropriate set.
+   *
+   * If the import is local to this package then follow it by calling
+   * `recursivelyCollectDependencies()`.
+   *
+   * @returns `true` if the import was resolved (to an entry-point, a local import, or a
+   * deep-import), `false` otherwise.
+   */
+  protected processImport(
+      importPath: string, file: AbsoluteFsPath, dependencies: Set<AbsoluteFsPath>,
+      missing: Set<string>, deepImports: Set<string>, alreadySeen: Set<AbsoluteFsPath>): boolean {
+    const resolvedModule = this.moduleResolver.resolveModuleImport(importPath, file);
+    if (resolvedModule === null) {
+      return false;
+    }
+    if (resolvedModule instanceof ResolvedRelativeModule) {
+      const internalDependency = resolvedModule.modulePath;
+      if (!alreadySeen.has(internalDependency)) {
+        alreadySeen.add(internalDependency);
+        this.recursivelyCollectDependencies(
+            internalDependency, dependencies, missing, deepImports, alreadySeen);
+      }
+    } else if (resolvedModule instanceof ResolvedDeepImport) {
+      deepImports.add(resolvedModule.importPath);
+    } else {
+      dependencies.add(resolvedModule.entryPointPath);
+    }
+    return true;
+  }
 }
